Use jq to filter package.json down
Currently the upstream action installs every package in our bundle to
run Stylelint. This adds minutes to each linter run. We started to work
around this by only installing the packages we wanted, but it meant the
linter would install the most recent versions of packages vs. what we've
been using for local development (the versions in our `yarn.lock` file).

We can work around this and retain version locking by stripping out the
dependencies we don't want. This change uses `jq` to edit our
package.json file inside the action. It works like this:

1. iterate over the sections. If the section isn't `dependencies` or
   `devDependencies`, keep it as-is.
2. if the section contains dependencies, iterate over the keys of the
   section. Filter out any elements that don't contain the string
   `stylelint`.